clean type by  int64 -> int  uint64 -> uint

+++ fontmaker/core/math.go
@@ -1,10 +1,10 @@
 package core
 
-func Round(value float64) int64 {
+func Round(value float64) int {
 	if value < 0.0 {
 		value -= 0.5
 	} else {
 		value += 0.5
 	}
-	return int64(value)
+	return int(value)
 }
